Restyle thumbnailed images and galleries
The gaudy thick white borders are gone, both from the thumbnails and the
overlay!  The controls on the overlay now stay in consistent positions,
and the caption doesn't get fucked up on small screens.

Thumbnails are also now generated to a consistent /height/ rather than
width, which is slightly trickier design-wise but makes for a much more
pleasing appearance.