Annotate every code block with its language
Turns out about half of the total build time (10~12s!) was Pygments
trying to guess the language of unlabelled blocks, which I'm pretty sure
it does by just throwing every single lexer it knows about at the text
and picking the one that seems to fit the best.  It knows almost 600
lexers, so, that takes a while.